Fix Dispose method in RuntimeSquare
Added a check for the 'disposing' parameter to ensure that MyPen and MyBrush are only disposed of when appropriate. Without this check the following line in OnPaint event of RunTimeSquare was throwing an exception because MyBrush would be disposed by garbage collector.

e.Graphics.FillRectangle(MyBrush, new Rectangle(0, 0, Width - 1, Height - 1))